requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Pillow==2.7.0
bleach==1.4.1
alembic==0.7.4
psycopg2==2.6
requests==2.5.1
Flask-Admin==1.0.9
Flask-Cache==0.13.1
redis==2.10.3
lxml==3.4.2
celery==3.1.17
elasticsearch==1.4.0
pdfkit==0.5.0

git+https://github.com/Rikanishu/static-bundle.git
git+https://github.com/Rikanishu/flask-static-bundle.git
git+https://github.com/Rikanishu/flask-babel.git

# optional dependings
newspaper==0.0.9.6
langdetect==1.0.1

# for requests SNI
pyOpenSSL==0.14
ndg-httpsclient==0.3.3
